New branch

+++ src/services/users/user.service.ts
@@ -1,8 +1,9 @@
+import { StatusCodes } from "http-status-codes";
 import type {
 	CreateUserRequestDTO,
 	CreateuserResponseDTO,
 } from "../../dtos/users/users.dto";
-import { InvalidInputError } from "../../errors/users/notFoundUsers.error";
+import { AppError } from "../../errors/AppError";
 import type { UserRepositoryInterface } from "../../repositories/users/user.repository";
 
 export class UserService {
@@ -13,15 +14,23 @@ export class UserService {
 		return users;
 	}
 
+	async getById(id: number) {
+		const user = await this.userRepository.findById(id);
+		return user;
+	}
+
 	async create(data: CreateUserRequestDTO): Promise<CreateuserResponseDTO> {
 		const errors: string[] = [];
 		if (!data.name) errors.push("name");
 		if (!data.email) errors.push("email");
 
 		if (errors.length > 0) {
-			throw InvalidInputError({
-				details: errors,
-			});
+			throw new AppError(
+				"INVALID_INPUT",
+				"Missing required fields",
+				StatusCodes.BAD_REQUEST,
+				errors,
+			);
 		}
 
 		const newUser = await this.userRepository.create(data);
@@ -29,6 +38,14 @@ export class UserService {
 	}
 
 	async delete(id: number): Promise<void> {
+		const user = await this.userRepository.findById(id);
+		if (!user) {
+			throw new AppError(
+				"USER_NOT_FOUND",
+				"User not found",
+				StatusCodes.NOT_FOUND,
+			);
+		}
 		return this.userRepository.delete(id);
 	}
 }
